% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{cubeOptions}
\alias{cubeOptions}
\title{Options to control cube appearance.}
\usage{
cubeOptions(x_pos = 1, y_pos = 1, z_pos = 1, ...)
}
\arguments{
\item{x_pos}{initial position of x-axis slice.}

\item{y_pos}{initial position of y-axis slice.}

\item{z_pos}{initial position of z-axis slice.}

\item{...}{currently ignored.}
}
\description{
Options to control cube appearance.
}
