% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legendOptions}
\alias{legendOptions}
\title{Options to control legend appearance.}
\usage{
legendOptions(width = 200, height = 250, ...)
}
\arguments{
\item{width}{width of the legend in pixels.}

\item{height}{height of the legend in pixels.}

\item{...}{currently ignored.}
}
\description{
Options to control legend appearance.
}
